# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Jeremiah Generation Youth API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /profiles/me:
    get:
      operationId: getMyProfile
      tags: [profiles]
      summary: Get current user profile
      responses:
        "200":
          description: Profile found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
        "404":
          description: Profile not found
    patch:
      operationId: updateMyProfile
      tags: [profiles]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profiles/register:
    post:
      operationId: registerVisitor
      tags: [profiles]
      summary: Register as a first-time visitor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VisitorRegistration"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profiles:
    get:
      operationId: listProfiles
      tags: [profiles]
      summary: List all profiles (leader/admin only)
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [super_admin, leader, member, visitor]
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of profiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Profile"
  /profiles/{id}:
    get:
      operationId: getProfile
      tags: [profiles]
      summary: Get a profile by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profiles/{id}/promote:
    post:
      operationId: promoteToMember
      tags: [profiles]
      summary: Promote visitor to member (leader/admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Promoted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"
  /profiles/{id}/revoke-membership:
    post:
      operationId: revokeMembership
      tags: [profiles]
      summary: Revoke membership (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Profile"

  # Events
...
  /rsvps/event/{eventId}:
    get:
      operationId: listEventRsvps
      tags: [rsvps]
      summary: List RSVPs for a specific event (leader/admin only)
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: RSVPs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rsvp"
  /rsvps/{eventId}:
    post:
      operationId: upsertRsvp
      tags: [rsvps]
      summary: RSVP to an event (creates or updates)
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RsvpInput"
      responses:
        "200":
          description: RSVP saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rsvp"

  # Membership Requests
...
  /qrcodes/regenerate:
    post:
      operationId: regenerateQrCode
      tags: [qrcodes]
      summary: Regenerate a QR code (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QrCodeRegenerateInput"
      responses:
        "200":
          description: New QR code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCode"
  /qrcodes/{slug}:
    get:
      operationId: resolveQrSlug
      tags: [qrcodes]
      summary: Resolve QR slug to a redirect target
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrCodeResolution"
        "404":
          description: Slug not found or expired

  # Dashboard
...
components:
  schemas:
...
    Profile:
      type: object
      required: [id, full_name, role, created_at]
      properties:
        id:
          type: string
        clerk_id:
          type: ["string", "null"]
        full_name:
          type: string
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        gender:
          type: ["string", "null"]
          enum: [male, female, other, "null"]
        age:
          type: ["integer", "null"]
        heard_from:
          type: ["string", "null"]
        role:
          type: string
          enum: [super_admin, leader, member, visitor]
        created_at:
          type: string
          format: date-time
    ProfileUpdate:
      type: object
      properties:
        full_name:
          type: string
        phone:
          type: string
        email:
          type: string
        gender:
          type: string
          enum: [male, female, other]
        age:
          type: integer
        heard_from:
          type: string
    VisitorRegistration:
      type: object
      required: [full_name, phone, gender, age, heard_from]
      properties:
        full_name:
          type: string
        phone:
          type: string
        email:
          type: ["string", "null"]
        gender:
          type: string
          enum: [male, female, other]
        age:
          type: integer
        heard_from:
          type: string
        clerk_id:
          type: ["string", "null"]
    Event:
      type: object
      required: [id, title, date, time, location, is_public, created_at]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        date:
          type: string
          format: date
        time:
          type: string
        location:
          type: string
        created_by:
          type: ["string", "null"]
        age_min:
          type: ["integer", "null"]
        age_max:
          type: ["integer", "null"]
        custom_requirements:
          type: ["array", "null"]
          items:
            $ref: "#/components/schemas/EventRequirement"
        is_public:
          type: boolean
        rsvp_count:
          type: ["integer", "null"]
        attendance_count:
          type: ["integer", "null"]
        created_at:
          type: string
          format: date-time
    EventRequirement:
      type: object
      required: [label, required]
      properties:
        label:
          type: string
        required:
          type: boolean
...
    Rsvp:
      type: object
      required: [id, event_id, profile_id, status, created_at]
      properties:
        id:
          type: string
        event_id:
          type: string
        profile_id:
          type: string
        status:
          type: string
          enum: [going, not_going, maybe]
        created_at:
          type: string
          format: date-time
        event:
          $ref: "#/components/schemas/Event"
    RsvpInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [going, not_going, maybe]
...
    QrCode:
      type: object
      required: [id, slug, type, active, created_at]
      properties:
        id:
          type: string
        slug:
          type: string
        type:
          type: string
          enum: [public, leader]
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
    QrCodeRegenerateInput:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [public, leader]
    QrCodeResolution:
      type: object
      required: [slug, type, redirect_to]
      properties:
        slug:
          type: string
        type:
          type: string
          enum: [public, leader]
        redirect_to:
          type: string
...