# lib/api-zod/src/generated/types/rsvp.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jeremiah Generation Youth API
 * OpenAPI spec version: 0.1.0
 */
import type { Event } from "./event";
import type { RsvpStatus } from "./rsvpStatus";

export interface Rsvp {
  id: string;
  event_id: string;
  profile_id: string;
  status: RsvpStatus;
  created_at: Date;
  event?: Event;
}