# lib/api-zod/src/generated/types/visitorRegistration.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Jeremiah Generation Youth API
 * OpenAPI spec version: 0.1.0
 */
import type { VisitorRegistrationGender } from "./visitorRegistrationGender";

export interface VisitorRegistration {
  full_name: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  gender: VisitorRegistrationGender;
  age: number;
  heard_from: string;
  /** @nullable */
  clerk_id?: string | null;
}